pyproject.toml
[project]
name = "grpc-seminar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
	"grpcio>=1.75.1",
	"grpcio-tools>=1.75.1",
]
